Fixed a space leak in `Stream.switchMap`.
The leak only happens in a specific degenerate usage scenario where the
`source` stream in `switchMap x2ys source` is effectively synchronous.
What happens then is that `switchMap` runs through the whole `source`
stream without producing output and this used to build up an unnecessary
chain of promises.

The fix required a number of *breaking changes* to the `Stream` module
signature, namely `Stream.joinWith`, `Stream.mapJoin` and
`Stream.Builder` had to be changed.  The changes make it possible to
eliminate some intermediate promise constructions.  It is unlikely that
these changes affect any users.