Wrap finalizer in try...finally
According to the CLR documentation, if the finalizer throws an exception, finally blocks may not be called. However, the C# compiler automatically wraps the body of finalizers in a try{} block, calling the base finalizer in finally{}.

Finalizer is used from deprecated members, so its use is also deprecated, but we'll keep it up to date until then.